Rename conponent
